back-end/cmps/requirements.txt
# Complaint Management Service (CMPS) Dependencies

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
motor==3.3.2
pymongo==4.6.1

# Data Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Environment & Configuration
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
